// frontend/src/openapi/models/TrainingExperiment.ts
export interface TrainingExperiment {
    /**
     *
     * @type {string}
     * @memberof TrainingExperiment
     */
    readonly id: string;
    /**
     *
     * @type {string}
     * @memberof TrainingExperiment
     */
    readonly project: string;
    /**
     *
     * @type {string}
     * @memberof TrainingExperiment
     */
    readonly name: string;
    /**
     *
     * @type {string}
     * @memberof TrainingExperiment
     */
    readonly purpose: string;
    /**
     *
     * @type {string}
     * @memberof TrainingExperiment
     */
    readonly requestKey: string;
    /**
     *
     * @type {any}
     * @memberof TrainingExperiment
     */
    readonly variants: any | null;
    /**
     *
     * @type {any}
     * @memberof TrainingExperiment
     */
    readonly protocol: any | null;
    /**
     *
     * @type {string}
     * @memberof TrainingExperiment
     */
    readonly evaluation: string | null;
    /**
     *
     * @type {string}
     * @memberof TrainingExperiment
     */
    readonly state: string;
    /**
     *
     * @type {{ [key: string]: any; }}
     * @memberof TrainingExperiment
     */
    readonly record: { [key: string]: any; };
    /**
     *
     * @type {Date}
     * @memberof TrainingExperiment
     */
    readonly createdAt: Date;
    /**
     *
     * @type {Date}
     * @memberof TrainingExperiment
     */
    readonly updatedAt: Date;
}